+++ packages/react-querybuilder/src/types/importExport.ts
@@ -44,11 +44,14 @@ export interface FormatQueryOptions {
    */
   ruleProcessor?: RuleProcessor;
   /**
-   * In the "sql"/"parameterized"/"parameterized_named" export formats,
-   * field names will be bracketed by this string. If an array of strings
+   * In the "sql", "parameterized", "parameterized_named", and "jsonata" export
+   * formats, field names will be bracketed by this string. If an array of strings
    * is passed, field names will be preceded by the first element and
-   * succeeded by the second element. A common value for this option is
-   * the backtick (```'`'```).
+   * succeeded by the second element.
+   *
+   * A common value for this option is the backtick (```'`'```).
+   *
+   * Tip: Use `fieldIdentifierSeparator` to bracket identifiers individually within field names.
    *
    * @default '' // the empty string
    *
@@ -61,6 +64,27 @@ export interface FormatQueryOptions {
    * // "[First name] = 'Steve'"
    */
   quoteFieldNamesWith?: string | [string, string];
+  /**
+   * When used in conjunction with the `quoteFieldNamesWith` option, field names will
+   * be split by this string, each part being individually processed as per the rules
+   * of the `quoteFieldNamesWith` configuration. The parts will then be re-joined
+   * by the same string.
+   *
+   * A common value for this option is `'.'`.
+   *
+   * A value of `false` will disable splitting/rejoining.
+   *
+   * @default false
+   *
+   * @example
+   * formatQuery(query, {
+   *   format: 'sql',
+   *   quoteFieldNamesWith: ['[', ']'],
+   *   fieldIdentifierSeparator: '.',
+   * })
+   * // "[dbo].[Musicians].[First name] = 'Steve'"
+   */
+  fieldIdentifierSeparator?: string;
   /**
    * Character to use for quoting string values in the SQL format.
    * @default `'`