+++ packages/react-querybuilder/src/utils/formatQuery/defaultRuleProcessorSQL.ts
@@ -1,6 +1,6 @@
 import type { RuleProcessor } from '../../types/index.noReact';
 import { defaultValueProcessorByRule } from './defaultValueProcessorByRule';
-import { mapSQLOperator, quoteFieldNamesWithArray } from './utils';
+import { mapSQLOperator, quoteFieldName } from './utils';
 
 /**
  * Default rule processor used by {@link formatQuery} for "sql" format.
@@ -10,6 +10,7 @@ export const defaultRuleProcessorSQL: RuleProcessor = (rule, opts) => {
     parseNumbers,
     escapeQuotes,
     quoteFieldNamesWith = ['', ''] as [string, string],
+    fieldIdentifierSeparator = '',
     quoteValuesWith = `'`,
     valueProcessor = defaultValueProcessorByRule,
     concatOperator = '||',
@@ -18,6 +19,7 @@ export const defaultRuleProcessorSQL: RuleProcessor = (rule, opts) => {
     parseNumbers,
     escapeQuotes,
     quoteFieldNamesWith,
+    fieldIdentifierSeparator,
     quoteValuesWith,
     concatOperator,
   });
@@ -34,7 +36,5 @@ export const defaultRuleProcessorSQL: RuleProcessor = (rule, opts) => {
     return '';
   }
 
-  const [qPre, qPost] = quoteFieldNamesWithArray(quoteFieldNamesWith);
-
-  return `${qPre}${rule.field}${qPost} ${operator} ${value}`.trim();
+  return `${quoteFieldName(rule.field, { quoteFieldNamesWith, fieldIdentifierSeparator })} ${operator} ${value}`.trim();
 };